sftp_packet_read: refactor
Now that sftp_packet_read doesn't call
_libssh2_channel_read_window_adjust explicitly, there is no need to
maintain a state variable.

The number of bytes read so far for the next packet is enough to
determine where to follow.

This patch simplifies the function logic also removing several of the
varibles in the sftp structure used to resume the function. Now, only
the partial_packet pointer and the number of bytes read so far are
used.