_libssh2_channel_receive_window_adjust: centralize adjustment calculation here
Probably due to the library evolution, window adjustment was not done
in some specific part of the code but scatered among several places.
Then the function _libssh2_channel_receive_window was used to actually
generate and send the adjustment packet or save the adjustment for
later if it was too small.

In this patch, calculation of the adjustment is moved inside
_libssh2_channel_receive_window_adjust, where it is done comparing the
initial window size and the current one plus the adjustment passed as
an argument (that now, is interpreted as the number of bytes that the
caller plans to read inmediately).

That ensures that the length of the buffered data is roughly the
initial window size given for the channel at most and keeps constant
for the full channel life.

The value of LIBSSH2_CHANNEL_MINADJUST has also been changed to 90Kb
which in practice, means sending and adjustment packet for every three
packets received.

Keeping the adjust_queue counter for the requests not being
inmediately honoured is not required any more (this is one of the
advantages of doing the calculation in
_libssh2_channel_receive_window_adjust). Also, the calling code is
simplified in several places.

Finally, as this patch changes the semantics of the adjustment
parameter it also changes the API functions
libssh2_channel_receive_window_adjust and
libssh2_channel_receive_window_adjust2, but due to the way the
calculation is done usage of these functions would just mean
increasing the window size for a while but them it would converge
again into the initial window size.