_libssh2_transport_read_all: add function
There were several places in the code where _libssh2_channel_read was
called in a loop in order to read all the packets available from the
network without blocking. But in every of those places it was done in
a slightly different ways and not all the error codes (specifically
LIBSSH2_ERROR_NONE) were handled correctly.

This patch factors that functionality into a new function and hides
the LIBSSH2_ERROR_NONE case as LIBSSH2_ERROR_EAGAIN.