libssh2_transport_send, send_existing: refactor those functions
The code calling LIBSSH2_SEND and handling errors and the
session_block_direction state was duplicated in the two functions in a
slightly different way.

Also, the way to call send_existing was quite convoluted and coupled
with libssh2_transport_send.

This patch redistributes the dutties of both functions in a much
simpler way and removes the duplicated code.