_libssh2_transport_read: handle packets with a bad MAC earlier
The MAC checking was being performed at libssh2_packet_add, requiring
dedicated variables in the session object to maintain the state and
passing the flag along the call chain.

Also, compressed data was processed before calling libssh2_packet_add
so, that error may well become masked as a decompression error.

Now, the bad MAC error is handled and reported as soon as it is
detected.

There is a minor backguard incompatibility being introduced here:
before this change, the data passed to the LIBSSH2_MACERROR callback
was the uncompressed data. Now, the raw payload is used.