_libssh2_packet_read: refactor packet reading and dispatching
Before this patch, in order to get data from the network
_libssh2_packet_read was called. It read a packet from the network and
then called _libssh2_packet_add to dispatch it. Inside
_libssh2_packet_add a key exchange could be initiated invoking
_libssh2_kex_exchange which again would call _libssh2_packet_read
again which was quite hairy as the state flags were not reentrant.

This patch reassigns dutties. Now the main function is
_libssh2_packet_read (mostly the former _libssh2_packet_add), which
calls _libssh2_transport_read to read decrypt and uncompress a packet.

The code is still hairy because of the reentrance thing, but I hope
that in this form it would be easier to do further simplifications.